requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorclass==2.2.2
cryptography==45.0.5
easygui==0.98.3
et_xmlfile==2.0.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.0
msoffcrypto-tool==5.4.2
olefile==0.47
oletools==0.60.2
openpyxl==3.1.5
pcodedmp==1.2.6
pillow==11.3.0
playwright==1.54.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
PyMuPDF==1.26.3
pyparsing==3.2.3
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
xlsxwriter==3.2.5
